// artifacts/auth/rules/enums.ts
export enum MethodEnum {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
  PATCH = 'PATCH',
  ALL = 'ALL',
  OPTIONS = 'OPTIONS',
  HEAD = 'HEAD',
  SEARCH = 'SEARCH',
}

export enum SCOPE_ALLOWED {
  GATEWAY_SERVICE = 'gateway_service',
  MERCHANT_SERVICE = 'merchant_service',
  AGENT_SERVICE = 'agent_service',
}